+++ bin/FastP_QC.py
@@ -38,16 +38,20 @@ def FastP_QC_before(input_json, output_file, name):
     raw_R1_bases = str(Info1['total_bases'])
     Q20_R1_bp = str(Info1['q20_bases'])
     Q20_R1_bp = str(Info1['q20_bases'])
-    Q20_R1_percent = str(Decimal(Info1['q20_bases']) / Decimal(Info1['total_bases']))
     Q30_R1_bp = str(Info1['q30_bases'])
-    Q30_R1_percent = str(Decimal(Info1['q30_bases']) / Decimal(Info1['total_bases']))
     Info2 = data['read2_before_filtering']
+        #check if there are reads otherwise dividing by zero throw
+    if Decimal(Info1['total_bases']) == 0:
+        Q20_R1_percent, Q20_R2_percent, Q30_R1_percent, Q30_R2_percent = (str(0) for i in range(4))
+    else:
+        Q20_R1_percent = str(Decimal(Info1['q20_bases']) / Decimal(Info1['total_bases']))
+        Q30_R1_percent = str(Decimal(Info1['q30_bases']) / Decimal(Info1['total_bases']))
+        Q20_R2_percent = str(Decimal(Info2['q20_bases']) / Decimal(Info2['total_bases']))
+        Q30_R2_percent = str(Decimal(Info2['q30_bases']) / Decimal(Info2['total_bases']))
     raw_R2_reads = str(Info2['total_reads'])
     raw_R2_bases = str(Info2['total_bases'])
     Q20_R2_bp = str(Info2['q20_bases'])
-    Q20_R2_percent = str(Decimal(Info2['q20_bases']) / Decimal(Info2['total_bases']))
     Q30_R2_bp = str(Info2['q30_bases'])
-    Q30_R2_percent = str(Decimal(Info2['q30_bases']) / Decimal(Info2['total_bases']))
     Line = name + '\t' + raw_R1_reads + '\t' + raw_R1_bases + '\t' + raw_R2_reads + '\t' + raw_R2_bases + '\t' + Q20_Total + '\t' + Q30_Total + '\t' + Q20_R1_bp + '\t' + Q20_R2_bp + '\t' + Q20_R1_percent + '\t' + Q20_R2_percent + '\t' + Q30_R1_bp + '\t' + Q30_R2_bp + '\t' + Q30_R1_percent + '\t' + Q30_R2_percent + '\t' + Total_Sequenced_bp + '\t' + Total_Sequenced_reads
     Out.write(Line)
     Out.close()
@@ -68,16 +72,20 @@ def FastP_QC_after(input_trimmed_json, input_singles_json, output_file, name):
     raw_R1_reads = str(Info1['total_reads'])
     raw_R1_bases = str(Info1['total_bases'])
     Q20_R1_bp = str(Info1['q20_bases'])
-    Q20_R1_percent = str(Decimal(Info1['q20_bases']) / Decimal(Info1['total_bases']))
     Q30_R1_bp = str(Info1['q30_bases'])
-    Q30_R1_percent = str(Decimal(Info1['q30_bases']) / Decimal(Info1['total_bases']))
     Info2 = data['read2_after_filtering']
+    #check if there are reads otherwise dividing by zero throw
+    if Decimal(Info1['total_bases']) == 0:
+        Q20_R1_percent, Q20_R2_percent, Q30_R1_percent, Q30_R2_percent = (str(0) for i in range(4))
+    else:
+        Q20_R1_percent = str(Decimal(Info1['q20_bases']) / Decimal(Info1['total_bases']))
+        Q30_R1_percent = str(Decimal(Info1['q30_bases']) / Decimal(Info1['total_bases']))
+        Q20_R2_percent = str(Decimal(Info2['q20_bases']) / Decimal(Info2['total_bases']))
+        Q30_R2_percent = str(Decimal(Info2['q30_bases']) / Decimal(Info2['total_bases']))
     raw_R2_reads = str(Info2['total_reads'])
     raw_R2_bases = str(Info2['total_bases'])
     Q20_R2_bp = str(Info2['q20_bases'])
-    Q20_R2_percent = str(Decimal(Info2['q20_bases']) / Decimal(Info2['total_bases']))
     Q30_R2_bp = str(Info2['q30_bases'])
-    Q30_R2_percent = str(Decimal(Info2['q30_bases']) / Decimal(Info2['total_bases']))
     f = open(input_singles_json)
     data = json.load(f)
     f.close()