+++ bin/fairy_proc.sh
@@ -55,7 +55,7 @@ read=$(echo "${full_name}" | grep -oP "_R\d_" | cut -f2 -d"_")
 
 if grep -q -e "error" -e "unexpected" ${prefix}.txt; then
 	#prefix=${prefix%%_*}
-	echo "FAILED CORRUPTION CHECK! CANNOT UNZIP FASTQ FILE. CHECK FASTQ FILE ${prefix}_${read} FOR CORRUPTION!" >> ${prefix}_results.txt
+	echo "FAILED CORRUPTION CHECK! CANNOT UNZIP FASTQ FILE. CHECK FASTQ FILE ${prefix}_${read} FOR CORRUPTION!" >> ${prefix}_summary.txt
 	
 	#error warning for line_summary channel
 	echo "ID	Auto_QC_Outcome	Warning_Count	Estimated_Coverage	Genome_Length	Assembly_Ratio_(STDev)	#_of_Scaffolds_>500bp	GC_%	Species	Taxa_Confidence	Taxa_Coverage	Taxa_Source	Kraken2_Trimd	Kraken2_Weighted	MLST_Scheme_1	MLST_1	MLST_Scheme_2	MLST_2	GAMMA_Beta_Lactam_Resistance_Genes	GAMMA_Other_AR_Genes	AMRFinder_Point_Mutations	Hypervirulence_Genes	Plasmid_Incompatibility_Replicons	Auto_QC_Failure_Reason" > ${prefix}_summaryline_failure.tsv
@@ -130,5 +130,5 @@ if grep -q -e "error" -e "unexpected" ${prefix}.txt; then
 	echo "WARNINGS: out of line with what is expected and MAY cause problems downstream."  >> "${sample_name}.synopsis"
 	echo "ALERT: something to note, does not mean it is a poor-quality assembly."  >> "${sample_name}.synopsis"
 else
-	echo "PASSED: File ${prefix}_${read} is not corrupt." >> ${prefix}_results.txt
+	echo "PASSED: File ${prefix}_${read} is not corrupt." >> ${prefix}_summary.txt
 fi