+++ CHANGELOG.md
@@ -189,9 +189,12 @@ Below are the list of changes to phx since is initial release. As fixes can take
 - `GRiPHin_Summary.xlsx` was highlighting coverage outside 40-100x despite `--coverage`` setting, changes made to respect `--coverage` flag.  
 - Added a fix to handle when auto select by the mlst script chooses the wrong taxonomy. PHX will force a rerun in cases where the taxonomy is known but initial mlst is run against incorrect scheme. Known instances found so far include: *E. coli* (Pasteur) being incorrectly indentified as *Aeromonas* and *E. coli* (Pasteur) being identified as *Klebsiella*. The scoring in the MLST program was updated and can now cause lower count perfect hits (e.g. 6 of 6 *Aeromonas* genes at 100%) to be scored higher than novel correct hits (e.g. 7 of 8 at 100%, 1 novel gene).  
 - Corrected instance where, in some cases, an mlst scheme could not be determined that a proper out file was not created.  
-- Updated FAIry to catch additional file corruption errors.
-   - FAIry will detect file corruption that prevents the completion of gzip and zcat and generate a synopsis file when needed.
-   - FAIry will detect and report out if R1/R2 fastqs do not have equal number of reads in the files.
+- Fixed issue where samples that failed SPAdes did not have `--coverage` parameter respected when generating synopsis file.  
+- Updated FAIry( file integrity check) to catch additional file integrity errors.
+   - FAIry detects and reports when:  
+      - Corrupt fastq files that prevents the completion of gzip and zcat and generate a synopsis file when needed.  
+      - If R1/R2 fastqs that do not have equal number of reads in the files.  
+      - If there are no reads or scaffolds left after filtering and read trimming steps, respectively.
 
 **Container Updates:**  
 - Base container bummped up to v2.1.0